Solved 14

+++ src/_002/_002.java
@@ -11,9 +11,9 @@ public class _002 {
 	}
 
 	public _002(int max) {
-		List<Integer> seq = Sequence.fibonacciMaxTerm(max);
+		List<Long> seq = Sequence.fibonacciMaxTerm(max);
 		int sum = 0;
-		for(Integer i : seq) {
+		for(Long i : seq) {
 			if(i % 2 == 0) {
 				sum += i;
 			}
